Adicionando dockerfile, jenkinsfile, readme, chromedriver para o Linux e ajuste das dependencias do docker

+++ src/test/java/Setup/TestRule.java
@@ -1,54 +1,59 @@
 package Setup;
 
+import Utils.PropertyReader;
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
-import utils.PropertiesReader;
-
 
 public class TestRule {
 
-	protected static WebDriver driver;
+	private static WebDriver driver;
+	private static String urlSimulador;
 
 	@Before
-	public void beforeCenario(Scenario s){
+	public void beforeCenario(Scenario cenario){
 
-		ChromeOptions options = new ChromeOptions();
-        String headless = "FALSE";
+		PropertyReader propertyReader = new PropertyReader();
 
-		String pathProjeto = System.getProperty("user.dir");
-		String sistemaOperacional =  System.getProperty("os.name");
+		if (cenario.getSourceTagNames().contains("@UI")) {
 
-		if (sistemaOperacional.toUpperCase().contains("WINDOWS")) {
-			System.setProperty("webdriver.chrome.driver", pathProjeto + "/drivers/chromedriver.exe");
-			System.setProperty("os.download.path", System.getProperty("user.home").replaceAll("\\\\", "/") + "/Downloads/");
-		} else {
-			System.setProperty("webdriver.chrome.driver", pathProjeto + "/drivers/chromedriver");
-			System.setProperty("os.download.path", System.getProperty("user.home") + "/Downloads/");
+			urlSimulador = propertyReader.getProperty("url_simulador");
+			ChromeOptions options = new ChromeOptions();
+			String pathProjeto = System.getProperty("user.dir");
+
+			if (System.getProperty("os.name").toUpperCase().contains("WINDOWS")) {
+				System.setProperty("webdriver.chrome.driver", pathProjeto + "/drivers/chromedriver.exe");
+				System.setProperty("os.download.path", System.getProperty("user.home")
+						.replaceAll("\\\\", "/") + "/Downloads/");
+			} else {
+				System.setProperty("webdriver.chrome.driver", pathProjeto + "/drivers/chromedriver");
+				System.setProperty("os.download.path", System.getProperty("user.home") + "/Downloads/");
+			}
+
+			if (Boolean.parseBoolean(propertyReader.getProperty("headless"))) options.addArguments("headless");
+
+			driver = new ChromeDriver(options);
+			driver.manage().deleteAllCookies();
+
+			if (!Boolean.parseBoolean(propertyReader.getProperty("headless")))	driver.manage().window().maximize();
+
 		}
-
-		if(headless.equals("TRUE")) {
-			options.addArguments("headless");
-			driver.manage().window().maximize();
-		}
-		options.addArguments("--incognito");
-
-		driver = new ChromeDriver(options);
-		driver.manage().deleteAllCookies();
-
-
 	}
 	
 	public static WebDriver getDriver() {
 		return driver;
 	}
-	
+
+	public static String getUrlSimulador() {
+		return urlSimulador;
+	}
+
 	@After
 	public void afterCenario(Scenario cenario) {
-		driver.quit();
+		if (cenario.getSourceTagNames().contains("@UI")) driver.quit();
 	}
 }
